Unify seed data.
